requirements.txt
appdirs==1.4.4
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.6
configparser==5.3.0
crayons==0.4.0
cryptography==39.0.0
cssselect==1.2.0
dateparser==1.1.7
demjson3==3.0.6
Django==4.1.6
django-crispy-forms==1.14.0
enum34==1.1.6
et-xmlfile==1.0.1
facebook-scraper==0.2.59
fake-headers==1.0.2
fake-useragent==1.1.1
html5lib==1.1
idna==2.9
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipaddress==1.0.17
jdcal==1.4.1
keyring==10.6.0
keyrings.alt==3.0
linkedin-api==2.0.0a5
lxml==4.5.0
numpy==1.19.0
openpyxl==3.0.4
pandas==1.0.5
parse==1.19.0
Pillow==9.4.0
pycparser==2.21
pycryptodome==3.17
pyee==8.2.2
PyMySQL==1.0.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
pyxdg==0.25
regex==2022.10.31
requests==2.23.0
requests-html==0.10.0
SecretStorage==2.3.1
secure-smtplib==0.1.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0
sqlparse==0.4.3
threaded==1.0.0
thready==0.1.5
tqdm==4.64.1
typing==3.6.4
tzdata==2022.7
tzlocal==4.2
urllib3==1.25.9
w3lib==2.1.1
webdriver-manager==3.2.2
webencodings==0.5.1
websockets==10.4
wincertstore==0.2
XlsxWriter==1.0.2
zipfile36==0.1.3
zipp==3.12.0
whitenoise
